backend/requirements.txt
Django==5.0.1
django-solo==2.2.0
djangorestframework==3.14.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
psycopg2==2.9.9
pillow==10.2.0